Renumber the index migration and give the mappers their own file
Main took a higher migration number while this branch was in flight, so the
partial-index migration is re-stamped above it: a version that sorts below
origin/main's highest applies in a different place on a fresh database than on
one already past it, and anything order-dependent then leaves two schemas that
differ with nothing to report it.

tags.go crossed the 500-line ceiling, so the two wire mappers move to the
transport file beside the handlers that call them. The usage counter also
carried a table parameter that was always the entity type spelled twice.
